Breaking change: rename package eventmux to event
Release: v0.4.0 x/eventlog/v0.3.0

Also: rename New to NewMux.

This paves the way for the introduction of alternative brokers.
