+++ event/doc.go
@@ -1,0 +1,2 @@
+// Package event provides in-memory event brokers.
+package event
