+++ event/none_test.go
@@ -1,13 +1,13 @@
-package eventmux_test
+package event_test
 
 import (
-	"artk.dev/eventmux"
+	"artk.dev/event"
 	"context"
 	"testing"
 )
 
 func TestNone(t *testing.T) {
-	err := eventmux.None[Event](context.TODO(), exampleEvent())
+	err := event.None[Event](context.TODO(), exampleEvent())
 	if err != nil {
 		t.Error("unexpected error:", err)
 	}
