--- eventmux/doc.go
@@ -1,2 +1,0 @@
-// Package eventmux provides an in-memory event multiplexer.
-package eventmux
