+++ x/eventlog/go.mod
@@ -3,7 +3,7 @@ module artk.dev/x/eventlog
 go 1.22.0
 
 require (
-	artk.dev v0.3.0
+	artk.dev v0.4.0
 	artk.dev/x/testlog v0.2.0
 )
